+++ packages/common/driver-definitions/api-report/driver-definitions.legacy.beta.api.md
@@ -734,6 +734,9 @@ export enum MessageType {
     SummaryNack = "summaryNack"
 }
 
+// @beta @input
+export type MinimumVersionForCollab = `${1 | 2}.${bigint}.${bigint}` | `${1 | 2}.${bigint}.${bigint}-${string}`;
+
 // @beta @legacy
 export enum NackErrorType {
     BadRequestError = "BadRequestError",
