+++ packages/drivers/local-driver/src/localService.ts
@@ -21,11 +21,11 @@ import type {
 	FluidContainerWithService,
 	Registry,
 	ServiceClient,
+	ServiceOptions,
 } from "@fluidframework/driver-definitions/internal";
 import {
 	type ContainerRuntimeLoader,
 	type ContainerRuntimeLoaderParams,
-	type ServiceOptions,
 	makeCodeLoader,
 	makeServiceClientImpl,
 	rootDataStoreId,
