+++ packages/framework/fluid-framework/src/index.ts
@@ -101,8 +101,6 @@ export type {
 	SharedObjectKindAlpha,
 	DataStoreContext,
 } from "@fluidframework/shared-object-base/internal";
-export type { MinimumVersionForCollab } from "@fluidframework/runtime-definitions/internal";
-export type { ServiceOptions } from "@fluidframework/runtime-utils/internal";
 export type {
 	DataStoreCreator,
 	DataStoreKey,
@@ -111,9 +109,11 @@ export type {
 	FluidContainer,
 	FluidContainerAttached,
 	FluidContainerWithService,
+	MinimumVersionForCollab,
 	Registry,
 	RegistryKey,
 	ServiceClient,
+	ServiceOptions,
 } from "@fluidframework/driver-definitions/internal";
 export {
 	basicKey,
