+++ packages/runtime/runtime-utils/src/index.ts
@@ -82,7 +82,6 @@ export type {
 	MinimumMinorSemanticVersion,
 	SemanticVersion,
 } from "./compatibilityBase.js";
-export type { ServiceOptions } from "./serviceClientBase.js";
 export type { Audience } from "./serviceClientBase.js";
 export {
 	DataStoreKindImplementation,
